Day79 LLM Studio/requirements.txt
textual>=0.47.0
rich>=13.7.0
llama-cpp-python>=0.2.50
fastapi>=0.109.0
uvicorn>=0.27.0
huggingface-hub>=0.20.0
pydantic>=2.5.0
httpx>=0.26.0
pyyaml>=6.0.1
